Batching : Instead of relying on UCTSearch turn batching on/off, have OpenCLScheduler predict what it needs to do
This makes code simpler.
We have no risk of bugs that can be caused by missing a corner case where we need to disable batching.
Still we can mostly fill our batches without much loss.

+++ src/Leela.cpp
@@ -252,7 +252,7 @@ static void parse_commandline(int argc, char *argv[]) {
             num_threads = cfg_max_threads;
         }
 #ifdef USE_OPENCL
-        if (num_threads < cfg_batch_size) {
+        if (static_cast<unsigned int>(num_threads) < cfg_batch_size) {
             printf("Threads number = %d must be larger than batch size = %d\n", num_threads, cfg_batch_size);
             exit(EXIT_FAILURE);
         }
